refactor: adjust structure of trailMap file tree
